fix Vue routing to separate slide editor from metadata editor
